fruit-img-rec/pyproject.toml
[tool.poetry]
name = "fruit-img-rec"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.11"
quart = "^0.19.9"
torch = "^2.5.1"
pillow = "^11.0.0"
torchvision = "^0.20.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
